Handle request bodies appropriately. Streams are hard.
So what was happening is that getRawBody was being called, which
parsed the body into `req.text`. This had the side effect of consuming
the request's data stream, so when you later add `on('data')` or `pipe`
or anything like that, it is already empty. I had no idea where to look
for the text, and eventually found it by assuming the cause of the error
and tracing the request back to the source. (AKA, find all middleware being
called.)

Node.js streams have three different APIs that you can use, and they all conflict,
resulting in unexpected behavior. Here be dragons.